pyproject.toml
[project]
name = "taskmanagement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.2",
    "pytz>=2025.2",
]
